DesignMatch Producto Final/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.35
werkzeug==3.0.1
psycopg2-binary==2.9.9
